backend/requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
SQLAlchemy==2.0.32
psycopg[binary]==3.2.13
python-jose==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
